+++ packages/melonjs/src/application/header.ts
@@ -1,0 +1,39 @@
+import * as device from "../system/device";
+import type Application from "./application.ts";
+
+/**
+ * display information
+ * @param app - the game application instance calling this function
+ */
+export function consoleHeader(app: Application): void {
+	const renderType = app.renderer.type;
+	const gpu_renderer =
+		typeof (app.renderer as any).GPURenderer === "string"
+			? ` (${(app.renderer as any).GPURenderer})`
+			: "";
+	const depthTesting =
+		renderType.includes("WebGL") &&
+		(app.renderer as any).depthTest === "z-buffer"
+			? "Depth Test | "
+			: "";
+	const audioType = device.hasWebAudio ? "Web Audio" : "HTML5 Audio";
+
+	// output video information in the console
+	console.log(
+		`${renderType} renderer${gpu_renderer} | ${depthTesting}${audioType} | ` +
+			`pixel ratio ${device.devicePixelRatio} | ${
+				device.platform.nodeJS
+					? "node.js"
+					: device.platform.isMobile
+						? "mobile"
+						: "desktop"
+			} | ${device.getScreenOrientation()} | ${device.language}`,
+	);
+
+	console.log(
+		`resolution: ` +
+			`requested ${(app.settings as any).width}x${
+				(app.settings as any).height
+			}, got ${app.renderer.width}x${app.renderer.height}`,
+	);
+}