+++ packages/melonjs/src/input/input.ts
@@ -4,15 +4,11 @@
 
 /**
  * specify if melonJS should prevent all default browser action on registered events.
- * @public
- * @type {boolean}
  * @default true
- * @name preventDefault
- * @memberof input
  */
-export let preventDefault = true;
+export const preventDefault: boolean = true;
 
-export * from "./gamepad.js";
+export * from "./gamepad.ts";
 export * from "./key.ts";
 export * from "./keyboard.ts";
-export * from "./pointerevent.js";
+export * from "./pointerevent.ts";